tests/data:
Test files for exercising pre-calculation filtering of sources when generating
task work.

In this test job, there is a single site of interest for the computation and 2
point sources. Once the point sources are dilated to the integration distance
(maximum_distance) of 180 km, one of them intersects with the point of
interest, and the other does not.

This way we can clearly test that filtering is working and that we only
generate tasks/work for thos sources in the model which are close enough.
Side note: The tasks/work are determined by the records inserted into
`source_progress`.